+++ examples/cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @workglow/cli
 
+## 0.0.89
+
+### Patch Changes
+
+- Fix subgraph reactive
+- Updated dependencies
+  - @workglow/test@0.0.89
+  - @workglow/ai@0.0.89
+  - @workglow/ai-provider@0.0.89
+  - @workglow/storage@0.0.89
+  - @workglow/task-graph@0.0.89
+  - @workglow/tasks@0.0.89
+  - @workglow/util@0.0.89
+
 ## 0.0.88
 
 ### Patch Changes