+++ packages/task-graph/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @workglow/task-graph
 
+## 0.0.89
+
+### Patch Changes
+
+- Fix subgraph reactive
+- Updated dependencies
+  - @workglow/job-queue@0.0.89
+  - @workglow/storage@0.0.89
+  - @workglow/util@0.0.89
+
 ## 0.0.88
 
 ### Patch Changes