+++ packages/util/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @workglow/util
 
+## 0.0.89
+
+### Patch Changes
+
+- Fix subgraph reactive
+
 ## 0.0.88
 
 ### Patch Changes